fix(picker): constrain TUI-only helper tests to the !lazy_fzf build
fuzzyMatch, sessionStateIcon and truncateString live in files tagged
//go:build !lazy_fzf, but their tests sat in the untagged sort_test.go.
That compiled under the default build (so make test and CI passed) yet
broke go test -tags lazy_fzf ./..., which the goreleaser before-hook runs
-- failing every release since the refactor. Move those tests into a
!lazy_fzf-tagged file so both build variants compile.
